pyproject.toml
[project]
name = "voproshalych"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
 dependencies = [
    "python-dotenv",
    "aiohttp",
    "langchain",
    "langchain-community",
    "atlassian-python-api",
    "beautifulsoup4",
    "lxml",
    "pypdf",
    "sentence-transformers",
    "sqlalchemy",
    "psycopg2-binary",
    "pgvector",
    "pytest",
    "alembic>=1.18.4",
    "gradio>=6.5.1",
    "openai",
    "rouge-score",
    "sacrebleu",
    "scikit-learn",
    "pandas",
    "plotly",
    "umap-learn",
]

[tool.black]
line-length = 88
target-version = ['py312']

[dependency-groups]
dev = [
    "gradio>=6.5.1",
]
